Add `TupperIntervalSet.recip`/`sin_over_x`.
